# modules/fbs-core/api/api-docs.yml
openapi: 3.0.0
info:
  title: Feedbacksystem
  description: Feedbacksystem API
  version: 1.0.0
  contact:
    email: Andrej.Sajenko@mni.thm.de
  license:
    name: CC BY-NC-SA 4.0
    url: https://creativecommons.org/licenses/by-nc-sa/4.0/
...
security:
  - JWT: [ ]
paths:
...
  /courses:
    get:
      operationId: getCourses

      tags:
        - Course
      summary: Get a list of all courses
      description: ""
      parameters:
        - name: visible
          in: query
          description: Returns all courses that satisfy the visibility value [true, false]
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          headers:
            Authorization:
              description: Bearer
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
    post:
      operationId: createCourse

      tags:
        - Course
      summary: Create a new course
      description: ""
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Course"
        description: Course object to create
        required: true
      responses:
        "201":
          description: Created
          headers:
            Authorization:
              description: Bearer
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
        "400":
          description: Invalid Format
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
...
  "/courses/{cid}/tasks":
    get:
      operationId: getCourseTasks

      tags:
        - Task
      summary: Get all tasks of a course
      parameters:
        - name: cid
          in: path
          description: course id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          headers:
            Authorization:
              description: Bearer
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
    post:
      operationId: createCourseTask

      tags:
        - Task
      summary: Create a new task for this course
      parameters:
        - name: cid
          in: path
          description: course id
          required: true
          schema:
            type: integer
      requestBody:
        $ref: "#/components/requestBodies/Task"
      responses:
        "201":
          description: Created
          headers:
            Authorization:
              description: Bearer
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
  "/courses/{cid}/tasks/results":
    get:
      operationId: getCourseTasksResults

      tags:
        - Task
      summary: Get all the result for the tasks
      parameters:
        - name: cid
          in: path
          description: course id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          headers:
            Authorization:
              description: Bearer
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserTaskResult"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
  "/courses/{cid}/tasks/{tid}":
    get:
      operationId: getCourseTask

      tags:
        - Task
      summary: Get a task by id
      parameters:
        - name: cid
          in: path
          description: course id
          required: true
          schema:
            type: integer
        - name: tid
          in: path
          description: task id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          headers:
            Authorization:
              description: Bearer
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
    put:
      operationId: updateCourseTask

      tags:
        - Task
      summary: Update an existing task for this course
      parameters:
        - name: cid
          in: path
          description: course id
          required: true
          schema:
            type: integer
        - name: tid
          in: path
          description: task id
          required: true
          schema:
            type: integer
      requestBody:
        $ref: "#/components/requestBodies/Task"
      responses:
        "200":
          description: OK
          headers:
            Authorization:
              description: Bearer
              schema:
                type: string
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Task"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
    delete:
      operationId: deleteCourseTask

      tags:
        - Task
      summary: Delete a task by id
      parameters:
        - name: cid
          in: path
          description: course id
          required: true
          schema:
            type: integer
        - name: tid
          in: path
          description: task id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          headers:
            Authorization:
              description: Bearer
              schema:
                type: string
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  "/courses/{cid}/tasks/{tid}/result":
    get:
      operationId: getCourseTaskResult

      tags:
        - Task
      summary: Gets the task result by id
      parameters:
        - name: cid
          in: path
          description: course id
          required: true
          schema:
            type: integer
        - name: tid
          in: path
          description: task id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          headers:
            Authorization:
              description: Bearer
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserTaskResult"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
...
  "/courses/{cid}/statistics/subtasks":
    get:
      operationId: getCourseSubtaskStatistic

      tags:
        - Course Statistics
      summary: Get the Subtask statistic for a course
      parameters:
        - name: cid
          in: path
          description: course id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          headers:
            Authorization:
              description: Bearer
              schema:
                type: string
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/SubTaskStatistics"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
...
servers:
  - url: https://feedback.mni.thm.de/api/v1
components:
  requestBodies:
    Task:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Task"
      description: Task object to create
      required: true
  securitySchemes:
    JWT:
      type: apiKey
      in: header
      name: Authorization
      description: Bearer
  schemas:
...
    Course:
      type: object
      required:
        - name
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        visible:
          type: boolean
        semesterId:
          type: integer
...
    Task:
      type: object
      required:
        - name
      properties:
        id:
          type: integer
        name:
          type: string
        isPrivate:
          type: boolean
        description:
          type: string
        deadline:
          type: string
          description: ISO-8601 time as string
        mediaType:
          type: string
          description: The media type for the submissions according to RFC 4288
        mediaInformation:
          description: Media type spezific Information
          oneOf:
            - $ref: "#/components/schemas/SpreadsheetInformationRequest"
            - $ref: "#/components/schemas/SpreadsheetInformationResponse"
        requirementType:
          type: string
          description: The type of requirment like "mandatory", "bonus/additional", etc.
          default: "mandatory"
          enum:
            - "mandatory"
            - "optional"
            - "practice"
        attempts:
          type: number
          nullable: true
    SpreadsheetInformationRequest:
      type: object
      properties:
        idField:
          type: string
        inputFields:
          type: string
        outputFields:
          type: string
        pointFields:
          type: string
    SpreadsheetInformationResponse:
      type: object
      properties:
        fields:
          type: object
          properties:
            inputs:
              type: array
              items:
                type: object
            outputs:
              type: array
              items:
                type: string
...
    UserTaskResult:
      type: object
      properties:
        points:
          type: number
        maxPoints:
          type: number
        passed:
          type: boolean
        submission:
          type: boolean
    SubTaskStatistics:
      type: object
      properties:
        taskID:
          type: number
        name:
          type: string
        subtasks:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              maxPoints:
                type: number
              avgPoints:
                type: number
...